Separate opt and deconv routines into separate files
