+++ pfb/deconv/__init__.py
@@ -8,4 +8,5 @@ date   - 31/03/2020
 from pfb.deconv.sara import sara
 from pfb.deconv.clean import clean
 from pfb.deconv.skoon import skoon
+from pfb.deconv.hogbom import hogbom
